@@ -48,7 +48,7 @@
         "npm-run-all": "4.1.5",
         "rimraf": "6.0.1",
         "ts-jest": "28.0.4",
-        "typescript": "4.6.3"
+        "typescript": "5.9.3"
     },
     "license-check-and-add-config": {
         "folder": "./src",
