@@ -71,7 +71,7 @@
     "tmp-promise": "3.0.2",
     "ts-jest": "28.0.4",
     "ts-node": "10.9.1",
-    "typescript": "4.6.3",
+    "typescript": "5.9.3",
     "webpack": "5.76.0",
     "webpack-cli": "4.9.1",
     "xregexp": "5.1.1",
