@@ -65,7 +65,7 @@
     "sinon-chai": "3.7.0",
     "ts-jest": "28.0.4",
     "ts-node": "10.9.1",
-    "typescript": "4.6.3",
+    "typescript": "5.9.3",
     "webpack-cli": "^6.0.1"
   },
   "dependencies": {
