@@ -41,7 +41,7 @@
         "license-check-and-add": "2.3.6",
         "npm-run-all": "4.1.5",
         "rimraf": "3.0.2",
-        "typescript": "4.6.3"
+        "typescript": "5.9.3"
     },
     "license-check-and-add-config": {
         "folder": "./src",
@@ -61,7 +61,8 @@
             ".tern-project",
             "./generated/concerto.metamodel@1.0.0.ts",
             "./generated/concerto.ts",
-            "./generated/concerto@1.0.0.ts"
+            "./generated/concerto@1.0.0.ts",
+            "./generated/concerto.decorator@1.0.0.ts"
         ],
         "file_type_method": "EXCLUDE",
         "file_types": [
