@@ -60,7 +60,7 @@
     "tmp-promise": "3.0.2",
     "ts-jest": "28.0.4",
     "ts-node": "10.9.1",
-    "typescript": "4.6.3"
+    "typescript": "5.9.3"
   },
   "dependencies": {
     "@accordproject/concerto-codegen": "^3.16.2",
