@@ -21,10 +21,8 @@ chai.should();
 chai.use(require('chai-things'));
 chai.use(require('chai-as-promised'));
 const sinon = require('sinon');
-const nock = require('nock');
 
-const { Agent, MockAgent, setGlobalDispatcher } = require('undici');
-const model = 'namespace org.accordproject.usa.business';
+const { Agent, setGlobalDispatcher } = require('undici');
 
 const defaultProcessFile = (name, data) => {
     return { name, data };
@@ -39,25 +37,10 @@ describe('HTTPModeFilelLoader', () => {
     }`;
 
     beforeEach(() => {
-        mockAgent = new MockAgent();
-
-        mockAgent
-            .get('https://raw.githubusercontent.com')
-            .intercept({ path: '/accordproject/business.cto' })
-            .reply(200, model);
-
-        mockAgent
-            .get('https://missing.com')
-            .intercept({ path: '/test' })
-            .reply(404);
-
-        setGlobalDispatcher(mockAgent);
-        mockAgent.disableNetConnect();
         sandbox = sinon.createSandbox();
     });
 
     afterEach(async () => {
-        await mockAgent.close();
         setGlobalDispatcher(new Agent());
         sandbox.restore();
     });
